django/aws_task/requirements.txt
asgiref==3.7.2
boto3==1.28.24
botocore==1.31.24
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
